requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
boto3==1.29.0
botocore==1.32.0
requests==2.31.0
click==8.1.7
pytest==7.4.3
pytest-cov==4.1.0
GitPython==3.1.40
pyyaml==6.0.1
checkov==3.2.30 